core: Expose a new `sign_tagged_message` method
This is the core of `sign_message` but allows arbitrary tags to be
used. This enables us to use the signer to create attestations
locally, that cannot be triggered by the node (whose requests will
always use the `Lightning Signed Message:` tag, and therefore be
distinct from the ones with other tags).

Changelog-Added: Added a `sign_tagged_message` method for raw signatures