scirps: update the nightly version
The PR [1] was updating the nightly feature, and this trigger some
update with the cargo.lock vesion.

So the build-all script was not able to compile anymore the stm32 crate.

This commit is fixing the nightly version used inside the script and
allow to compile the project again.

```
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 11.77s
+ cd lightning-storage-server
+ cargo build
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.07s
+ cd vls-signer-stm32
+ cargo +nightly-2022-11-15 build --features stm32f413,debug --release --bin demo_signer
error: failed to parse lock file at: /home/vincent/gitlab/work/lightning-signer/validating-lightning-signer/vls-signer-stm32/Cargo.lock

Caused by:
  lock file version `4` was found, but this version of Cargo does not understand this lock file, perhaps Cargo needs to be updated?
```

[1] https://gitlab.com/lightning-signer/validating-lightning-signer/-/merge_requests/712
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>