lssd: use official etcd image
bitnami image was migrated to bitnamilegacy namespace as they are 
planning to make there security hardened images a paid offering. We just
need an image for test so it’s fine to use this.

No using officially etcd image as that doesn’t setup a user for us and 
requires additional command line step using `etcdctl`